+++ apps/web/src/app/routes/router.tsx
@@ -1,6 +1,7 @@
 import { createBrowserRouter } from 'react-router-dom';
 
 import AppMainLayout from '@/app/layouts/AppMainLayout';
+import MobileLayout from '@/app/layouts/MobileLayout';
 import AgreementPage from '@/pages/agreement/AgreementPage';
 import HomePage from '@/pages/home/HomePage';
 import MapSearchPage from '@/pages/home/search/MapSearchPage';
@@ -21,7 +22,6 @@ import MonthlyRecordListPage from '@/pages/report/monthly-records/MonthlyRecordL
 import ReportDetailPage from '@/pages/report/ReportDetailPage';
 import ReportPage from '@/pages/report/ReportPage';
 import { ROUTE_PATHS, ROUTE_PATTERNS } from '@/shared/constants/routePaths';
-import MobileLayout from '@/shared/layout/MobileLayout';
 import PaddedLayout from '@/shared/layout/PaddedLayout';
 
 export const router = createBrowserRouter([
